refactor libse classes to psr 4 standard and load them via composer
