--- htdocs/libse/.htaccess
@@ -1,4 +1,0 @@
-<FilesMatch ".*">
-  Order deny,allow
-  Deny from all
-</FilesMatch>
