--- htdocs/libse/Cache/Where.php
@@ -1,19 +1,0 @@
-<?php
-
-/***************************************************************************
- *  For license information see doc/license.txt
- *
- *  Unicode Reminder メモ
- ***************************************************************************/
-class Cache_Where
-{
-    public static function active()
-    {
-        return ' caches.status = ' . Cache_Status::Active . ' AND (caches.date_activate IS NULL OR caches.date_activate <= NOW()) ';
-    }
-
-    public static function publishNow()
-    {
-        return ' caches.status = ' . Cache_Status::NotYetPubliced . ' AND caches.date_activate <= NOW() ';
-    }
-}
