--- htdocs/libse/Coordinate/Type.php
@@ -1,14 +1,0 @@
-<?php
-/***************************************************************************
- *  For license information see doc/license.txt
- *
- *  Unicode Reminder メモ
- ***************************************************************************/
-
-require_once(__DIR__ . '/../../lib2/logic/const.inc.php');
-
-class Coordinate_Type
-{
-    const ChildWaypoint = COORDINATE_WAYPOINT;
-    const UserNote = COORDINATE_USERNOTE;
-}
