--- htdocs/libse/Validator/AlwaysValid.php
@@ -1,14 +1,0 @@
-<?php
-
-/***************************************************************************
- *  For license information see doc/license.txt
- *
- *  Unicode Reminder メモ
- ***************************************************************************/
-class Validator_AlwaysValid
-{
-    public function isValid($value)
-    {
-        return true;
-    }
-}
