+++ htdocs/src/Oc/Libse/Cache/Where.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Oc\Libse\Cache;
+
+class WhereCache
+{
+    public static function active()
+    {
+        return ' caches.status = ' . StatusCache::Active . ' AND (caches.date_activate IS NULL OR caches.date_activate <= NOW()) ';
+    }
+
+    public static function publishNow()
+    {
+        return ' caches.status = ' . StatusCache::NotYetPubliced . ' AND caches.date_activate <= NOW() ';
+    }
+}
