+++ htdocs/src/Oc/Libse/Coordinate/Type.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Oc\Libse\Coordinate;
+
+/***************************************************************************
+ *  For license information see doc/license.txt
+ *
+ *  Unicode Reminder メモ
+ ***************************************************************************/
+
+require_once(__DIR__ . '/../../lib2/logic/const.inc.php');
+
+class TypeCoordinate
+{
+    const ChildWaypoint = COORDINATE_WAYPOINT;
+    const UserNote = COORDINATE_USERNOTE;
+}
