+++ htdocs/src/Oc/Libse/Validator/AlwaysValid.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Oc\Libse\Validator;
+
+/***************************************************************************
+ *  For license information see doc/license.txt
+ *
+ *  Unicode Reminder メモ
+ ***************************************************************************/
+class AlwaysValidValidator
+{
+    public function isValid($value)
+    {
+        return true;
+    }
+}
