+++ htdocs/src/Oc/Libse/Validator/Regex.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Oc\Libse\Validator;
+
+/***************************************************************************
+ *  For license information see doc/license.txt
+ *
+ *  Unicode Reminder メモ
+ ***************************************************************************/
+class RegexValidator
+{
+    private $regex;
+
+    public function __construct($regex)
+    {
+        $this->regex = $regex;
+    }
+
+    public function isValid($value)
+    {
+        return mb_ereg_match($this->regex, $value);
+    }
+}
